note: BruteForce/NM 정리
