+++ src/BruteForce/NM/BOJ15649.java
@@ -1,31 +1,31 @@
-package BruteForce;
+package BruteForce.NM;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.util.Arrays;
 import java.util.StringTokenizer;
 
-public class BOJ15655 {
+public class BOJ15649 {
 
     public static StringBuilder sb = new StringBuilder();
-    public static int N, M;
-    public static int base[];
+    public static int N,M;
+    public static int arr[];
     public static boolean visited[];
 
-    public static void permutationASC(int depth, int r){
+    public static void permutation(int depth){
         if(depth == M){
-            for(int i = 0; i<N; i++){
-                if(visited[i]) sb.append(base[i]).append(' ');
+            for(int val : arr){
+                sb.append(val).append(' ');
             }
             sb.append('\n');
             return;
         }
 
-        for(int i = r; i<N; i++){
+        for(int i = 0; i<N; i++){
             if(visited[i]) continue;
             visited[i] = true;
-            permutationASC(depth + 1, i + 1);
+            arr[depth] = i+1;
+            permutation(depth + 1);
             visited[i] = false;
         }
     }
@@ -36,20 +36,10 @@ public class BOJ15655 {
 
         N = Integer.parseInt(st.nextToken());
         M = Integer.parseInt(st.nextToken());
-
-        base = new int[N];
         visited = new boolean[N];
+        arr = new int[M];
 
-        st = new StringTokenizer(br.readLine());
-
-        for(int i = 0; i<N; i++){
-            base[i] = Integer.parseInt(st.nextToken());
-        }
-
-        Arrays.sort(base);
-
-        permutationASC(0, 0);
-
+        permutation(0);
         System.out.println(sb);
     }
 }
