+++ src/BruteForce/NM/BOJ15650.java
@@ -1,19 +1,32 @@
-package BruteForce;
+package BruteForce.NM;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.util.Arrays;
 import java.util.StringTokenizer;
 
-public class BOJ15656 {
+public class BOJ15650 {
 
     public static StringBuilder sb = new StringBuilder();
     public static int N,M;
-    public static int base[];
     public static int arr[];
 
-    public static void duplicatePermutation(int depth){
+    public static void useRecursive(int depth, int r){
+        if(depth == M){
+            for(int val : arr){
+                sb.append(val).append(' ');
+            }
+            sb.append('\n');
+            return;
+        }
+        if(r == N) return;
+
+        arr[depth] = r+1;
+        useRecursive(depth+1, r+1);
+        useRecursive(depth, r+1);
+    }
+
+    public static void permutationASC(int depth, int r){
         if(depth == M){
             for(int val : arr){
                 sb.append(val).append(' ');
@@ -22,9 +35,9 @@ public class BOJ15656 {
             return;
         }
 
-        for(int i = 0; i<N; i++){
-            arr[depth] = base[i];
-            duplicatePermutation(depth + 1);
+        for(int i = r; i<N; i++){
+            arr[depth] = i+1;
+            permutationASC(depth + 1, i+1);
         }
     }
 
@@ -34,17 +47,9 @@ public class BOJ15656 {
 
         N = Integer.parseInt(st.nextToken());
         M = Integer.parseInt(st.nextToken());
-
-        base = new int[N];
         arr = new int[M];
 
-        st = new StringTokenizer(br.readLine());
-        for(int i = 0; i<N; i++){
-            base[i] = Integer.parseInt(st.nextToken());
-        }
-
-        Arrays.sort(base);
-        duplicatePermutation(0);
+        permutationASC(0, 0);
         System.out.println(sb);
     }
 }
