+++ src/BruteForce/NM/BOJ15656.java
@@ -1,4 +1,4 @@
-package BruteForce;
+package BruteForce.NM;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,14 +6,14 @@ import java.io.InputStreamReader;
 import java.util.Arrays;
 import java.util.StringTokenizer;
 
-public class BOJ15657 {
+public class BOJ15656 {
 
     public static StringBuilder sb = new StringBuilder();
     public static int N,M;
-    public static int[] base;
-    public static int[] arr;
+    public static int arr[];
+    public static int base[];
 
-    public static void duplicatePermutationASC(int depth, int r){
+    public static void duplicatePermutation(int depth){
         if(depth == M){
             for(int val : arr){
                 sb.append(val).append(' ');
@@ -22,9 +22,9 @@ public class BOJ15657 {
             return;
         }
 
-        for(int i = r; i<N; i++){
+        for(int i = 0; i<N; i++){
             arr[depth] = base[i];
-            duplicatePermutationASC(depth+1, i);
+            duplicatePermutation(depth + 1);
         }
     }
 
@@ -34,9 +34,8 @@ public class BOJ15657 {
 
         N = Integer.parseInt(st.nextToken());
         M = Integer.parseInt(st.nextToken());
-
-        base = new int[N];
         arr = new int[M];
+        base = new int[N];
 
         st = new StringTokenizer(br.readLine());
         for(int i = 0; i<N; i++){
@@ -44,7 +43,7 @@ public class BOJ15657 {
         }
 
         Arrays.sort(base);
-        duplicatePermutationASC(0, 0);
+        duplicatePermutation(0);
         System.out.println(sb);
     }
 }
